Hook up snippet completion provider
